requirements.txt
Django==4.1
djangorestframework==3.13.1
drf-yasg==1.21.3
requests==2.28.1
